{{cookiecutter.project_slug}}/requirements.txt
-c constraints.txt
django==3.0.11
django-cors-headers
django-db-geventpool
django-debug-toolbar
django-environ
django-hijack
django-htmlmin
django-parler
django-sekizai
factory_boy
gevent
isort
pillow
psycopg2-binary
pytest-cov
pytest-django
python-memcached
snapshottest
splinter[django]
whitenoise
